+++ src/full/Agda/Syntax/Parser/Helpers.hs
@@ -32,7 +32,7 @@ import Agda.TypeChecking.Positivity.Occurrence
 import Agda.Utils.Either
 import Agda.Utils.Functor
 import Agda.Utils.Hash
-import Agda.Utils.List ( spanJust, chopWhen )
+import Agda.Utils.List ( spanJust, chopWhen, initLast )
 import Agda.Utils.List1 ( List1, pattern (:|), (<|) )
 import Agda.Utils.Monad
 import Agda.Utils.Null
@@ -88,23 +88,37 @@ figureOutTopLevelModule ds =
       -- We use the beginning of the file as beginning of the top level module.
       r = beginningOfFile $ getRange ds1
 
--- | Create a name from a string.
+-- | Create a name from a string. The boolean indicates whether a part
+-- of the name can be token 'constructor'.
+mkName' :: Bool -> (Interval, String) -> Parser Name
+mkName' constructor' (i, s) = do
+    let
+      xs = C.stringNameParts s
 
-mkName :: (Interval, String) -> Parser Name
-mkName (i, s) = do
-    let xs = C.stringNameParts s
-    mapM_ isValidId xs
+      -- The keyword constructor can appear as the only NamePart in the
+      -- last segment of a qualified name --- Foo.constructor refers to
+      -- the constructor of the record Foo.
+      constructor = case xs of
+        _ :| [] -> constructor'
+        _       -> False
+      --  The constructor' argument to mkName' determines whether this
+      --  is the last segment of a QName, the local variable constructor
+      --  additionally takes whether it's the only NamePart into
+      --  consideration.
+
+    mapM_ (isValidId constructor) xs
     unless (alternating xs) $ parseError $ "a name cannot contain two consecutive underscores"
     return $ Name (getRange i) InScope xs
     where
-        isValidId Hole   = return ()
-        isValidId (Id y) = do
+        isValidId _ Hole   = return ()
+        isValidId con (Id y) = do
           let x = rawNameToString y
               err = "in the name " ++ s ++ ", the part " ++ x ++ " is not valid"
           case parse defaultParseFlags [0] (lexer return) x of
             ParseOk _ TokId{}  -> return ()
             ParseFailed{}      -> parseError err
             ParseOk _ TokEOF{} -> parseError err
+            ParseOk _ (TokKeyword KwConstructor _) | con -> pure ()
             ParseOk _ t   -> parseError . ((err ++ " because it is ") ++) $ case t of
               TokQId{}      -> __IMPOSSIBLE__ -- "qualified"
               TokKeyword{}  -> "a keyword"
@@ -148,11 +162,17 @@ mkName (i, s) = do
         alternating (_    :| x   : xs) = alternating $ x :| xs
         alternating (_    :|       []) = True
 
+-- | Create a name from a string
+mkName :: (Interval, String) -> Parser Name
+mkName = mkName' False
+
 -- | Create a qualified name from a list of strings
 mkQName :: [(Interval, String)] -> Parser QName
-mkQName ss = do
-    xs <- mapM mkName ss
-    return $ foldr Qual (QName $ last xs) (init xs)
+mkQName ss | Just (ss0, ss1) <- initLast ss = do
+  xs0 <- mapM mkName ss0
+  xs1 <- mkName' True ss1
+  return $ foldr Qual (QName xs1) xs0
+mkQName _ = __IMPOSSIBLE__ -- The lexer never gives us an empty list of parts
 
 mkDomainFree_ :: (NamedArg Binder -> NamedArg Binder) -> Maybe Pattern -> Name -> NamedArg Binder
 mkDomainFree_ f p n = f $ defaultNamedArg $ Binder p $ mkBoundName_ n