+++ src/full/Agda/TypeChecking/Monad/Base.hs
@@ -4993,6 +4993,7 @@ data TypeError
         | GeneralizedVarInLetOpenedModule A.QName
         | MultipleFixityDecls (List1 (C.Name, Pair Fixity'))
         | MultiplePolarityPragmas (List1 C.Name)
+        | ConstructorNameOfNonRecord ResolvedName
     -- Concrete to Abstract errors
         | DeclarationsAfterTopLevelModule
         | IllegalDeclarationBeforeTopLevelModule