+++ CHANGELOG.md
@@ -63,6 +63,11 @@ Additions to the Agda syntax.
 
   See [#4275](https://github.com/agda/agda/issues/4275) for the proposal.
 
+* It is now always possible to refer to the name of a record type's
+  constructor, even if a name was not explicitly specified. This is done
+  using the new `(Record name).constructor` syntax; See [issue
+  #6964](https://github.com/agda/agda/issues/6964) for the motivation.
+
 Language
 --------